src/main/java/com/hr/management/test/AdminControllerTest.java
/*
 * package com.hr.management.test; import static org.junit.Assert.*; import
 * org.junit.Before; import org.junit.Test; import org.junit.runner.RunWith;
 * import org.mockito.InjectMocks; import org.mockito.Mock; import
 * org.mockito.MockitoAnnotations; import
 * org.springframework.http.ResponseEntity;
 * 
 * import com.hr.management.controller.AdminController; import
 * com.hr.management.service.AdminService;
 * 
 * @RunWith(MockitoJUnitRunner.class) public class AdminControllerTest {
 * 
 * @InjectMocks private AdminController adminController;
 * 
 * @Mock private AdminService adminService;
 * 
 * @Before public void init() { MockitoAnnotations.initMocks(this); }
 * 
 * @Test public void testCreateAdmin() { // Arrange AdminPayload adminPayload =
 * new AdminPayload(); // Create a sample payload
 * 
 * // Mock the behavior of adminService.createAdmin
 * when(adminService.createAdmin(any(AdminPayload.class)))
 * .thenReturn(ResponseEntity.ok().build());
 * 
 * // Act ResponseEntity<Object> response =
 * adminController.createAdmin(adminPayload);
 * 
 * // Assert assertEquals(200, response.getStatusCodeValue()); // Check if the
 * response status is OK }
 * 
 * @Test public void testGetAdminById() { // Arrange String adminId = "123"; //
 * Provide an existing admin ID
 * 
 * // Mock the behavior of adminService.getAdminById
 * when(adminService.getAdminById(adminId))
 * .thenReturn(ResponseEntity.ok().build());
 * 
 * // Act ResponseEntity<Object> response =
 * adminController.getAdminById(adminId);
 * 
 * // Assert assertEquals(200, response.getStatusCodeValue()); // Check if the
 * response status is OK }
 * 
 * // Similarly, create test methods for other controller endpoints
 * (updateAdmin, deleteAdmin, etc.) }
 */